Split up fam and rearrange for shared parts.
