+++ gnc/fam/package.xml
@@ -22,6 +22,7 @@
   <build_depend>ff_util</build_depend>
   <build_depend>config_reader</build_depend>
   <build_depend>gnc_autocode</build_depend>
+  <build_depend>pmc</build_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>nodelet</run_depend>
   <run_depend>ff_msgs</run_depend>
@@ -30,6 +31,7 @@
   <run_depend>ff_hw_msgs</run_depend>
   <run_depend>config_reader</run_depend>
   <run_depend>gnc_autocode</run_depend>
+  <run_depend>pmc</run_depend>
   <export>
     <nodelet plugin="${prefix}/nodelet_plugins.xml" />
   </export>
