+++ gnc/pmc/readme.md
@@ -1,0 +1,8 @@
+\page pmc Propulsion System
+
+This package includes two components revolving around Astrobee's propulsion system:
+
+* FAM (Force Allocation Module)--- Converts force and torque commands to nozzle opening angles
+* PMC simulator--- simulates the robot's propusion system. It is essentially the reverse of the
+  FAM--- given impeller commands and servo commands, it returns the force and torque applied to the robot.
+
