+++ simulation/package.xml
@@ -27,7 +27,7 @@
   <build_depend>ff_hw_msgs</build_depend>
   <build_depend>tf2_geometry_msgs</build_depend>
   <build_depend>ff_util</build_depend>
-  <build_depend>pmc_sim</build_depend>
+  <build_depend>pmc</build_depend>
   <build_depend>camera</build_depend>
   <build_depend>cv_bridge</build_depend>
   <build_depend>jsonloader</build_depend>
@@ -42,7 +42,7 @@
   <run_depend>ff_hw_msgs</run_depend>
   <run_depend>tf2_geometry_msgs</run_depend>
   <run_depend>ff_util</run_depend>
-  <run_depend>pmc_sim</run_depend>
+  <run_depend>pmc</run_depend>
   <run_depend>camera</run_depend>
   <run_depend>cv_bridge</run_depend>
   <run_depend>jsonloader</run_depend>
